feat(generators): install prettier with plugins
