+++ packages/nx-maven/src/generators/init/generator.ts
@@ -1,14 +1,9 @@
 import {
-  springBootVersion,
   kotlinVersion,
-  quarkusVersion,
   micronautVersion,
+  quarkusVersion,
+  springBootVersion,
 } from '@jnxplus/common';
-import {
-  addOrUpdateGitattributes,
-  addOrUpdatePrettierIgnore,
-  addOrUpdatePrettierRc,
-} from '../../utils/generators';
 import {
   ProjectConfiguration,
   Tree,
@@ -18,6 +13,7 @@ import {
   joinPathFragments,
   offsetFromRoot,
   updateJson,
+  writeJson,
 } from '@nx/devkit';
 import * as path from 'path';
 import { NxMavenGeneratorSchema } from './schema';
@@ -94,6 +90,7 @@ export default async function (tree: Tree, options: NxMavenGeneratorSchema) {
   addFiles(tree, normalizedOptions);
   updateNxJson(tree, normalizedOptions);
   updateGitIgnore(tree, options.skipWrapper);
+  installPrettier(tree);
   addOrUpdatePrettierRc(tree);
   addOrUpdatePrettierIgnore(tree);
   addOrUpdateGitattributes(tree);
@@ -110,7 +107,7 @@ export default async function (tree: Tree, options: NxMavenGeneratorSchema) {
   await formatFiles(tree);
 }
 
-export function updateNxJson(tree: Tree, options: NormalizedSchema) {
+function updateNxJson(tree: Tree, options: NormalizedSchema) {
   const plugin = {
     plugin: '@jnxplus/nx-maven',
     options: {
@@ -128,7 +125,7 @@ export function updateNxJson(tree: Tree, options: NormalizedSchema) {
   });
 }
 
-export function updateGitIgnore(tree: Tree, skipWrapper: boolean | undefined) {
+function updateGitIgnore(tree: Tree, skipWrapper: boolean | undefined) {
   const filePath = `.gitignore`;
   const contents = tree.read(filePath, 'utf-8') || '';
 
@@ -146,4 +143,68 @@ export function updateGitIgnore(tree: Tree, skipWrapper: boolean | undefined) {
 
   const newContents = contents.concat(mavenIgnore);
   tree.write(filePath, newContents);
+}
+
+function installPrettier(tree: Tree) {
+  updateJson(tree, 'package.json', (packageJson) => {
+    if (!packageJson.devDependencies['prettier']) {
+      packageJson.devDependencies['prettier'] = '^2.8.7';
+    }
+
+    if (!packageJson.devDependencies['@prettier/plugin-xml']) {
+      packageJson.devDependencies['@prettier/plugin-xml'] = '^2.2.0';
+    }
+
+    if (!packageJson.devDependencies['prettier-plugin-java']) {
+      packageJson.devDependencies['prettier-plugin-java'] = '^2.1.0';
+    }
+    return packageJson;
+  });
+}
+
+function addOrUpdatePrettierRc(tree: Tree) {
+  const prettierRcPath = `.prettierrc`;
+  if (tree.exists(prettierRcPath)) {
+    updateJson(tree, prettierRcPath, (prettierRcJson) => {
+      prettierRcJson.xmlWhitespaceSensitivity = 'ignore';
+      // return modified JSON object
+      return prettierRcJson;
+    });
+  } else {
+    writeJson(tree, prettierRcPath, {
+      xmlWhitespaceSensitivity: 'ignore',
+    });
+  }
+}
+
+function addOrUpdatePrettierIgnore(tree: Tree) {
+  const prettierIgnorePath = `.prettierignore`;
+  const mavenPrettierIgnore = '# Maven target\ntarget/';
+  if (tree.exists(prettierIgnorePath)) {
+    const prettierIgnoreOldContent =
+      tree.read(prettierIgnorePath, 'utf-8') || '';
+    const prettierIgnoreContent = prettierIgnoreOldContent.concat(
+      '\n',
+      mavenPrettierIgnore,
+    );
+    tree.write(prettierIgnorePath, prettierIgnoreContent);
+  } else {
+    tree.write(prettierIgnorePath, mavenPrettierIgnore);
+  }
+}
+
+function addOrUpdateGitattributes(tree: Tree) {
+  const gitattributesPath = `.gitattributes`;
+  const mavenWrapperGitattributes =
+    '# OS specific line endings for the Maven wrapper script\nmvnw text eol=lf\nmvnw.cmd text eol=crlf';
+  if (tree.exists(gitattributesPath)) {
+    const gitattributesOldContent = tree.read(gitattributesPath, 'utf-8') || '';
+    const gitattributesContent = gitattributesOldContent.concat(
+      '\n',
+      mavenWrapperGitattributes,
+    );
+    tree.write(gitattributesPath, gitattributesContent);
+  } else {
+    tree.write(gitattributesPath, mavenWrapperGitattributes);
+  }
 }
